+++ kintree_gui.py
@@ -1,10 +1,4 @@
-import flet as ft
-
-from kintree.gui.gui import kintree_gui
+from kintree.kintree_gui import main
 
 if __name__ == '__main__':
-    browser = False
-    if browser:
-        ft.app(target=kintree_gui, view=ft.WEB_BROWSER)
-    else:
-        ft.app(target=kintree_gui)
+    main()
